"Lista crear pedido terminada"

+++ app/src/main/java/isi/dam/sendmeal/Activities/CrearItem.java
@@ -54,7 +54,7 @@ public class CrearItem extends AppCompatActivity {
             }
             else{
                 Toast.makeText(getApplicationContext(), "Datos Registrados Exitosamente", Toast.LENGTH_SHORT).show();
-                Plato plato = new Plato(ingresoNombre.getEditText().getText().toString(), ingresoDescripcion.getEditText().getText().toString(), Float.valueOf(ingresoPrecio.getEditText().getText().toString()), Float.valueOf(ingresoCalorias.getEditText().getText().toString()));
+                Plato plato = new Plato(ingresoNombre.getEditText().getText().toString(), ingresoDescripcion.getEditText().getText().toString(), Double.valueOf(ingresoPrecio.getEditText().getText().toString()), Float.valueOf(ingresoCalorias.getEditText().getText().toString()));
                 Plato_repo.getInstance().crearPlato(plato,miHandler);
                 Toast.makeText(getApplicationContext(), plato.toString(), Toast.LENGTH_LONG).show();
             }
